// src/shared/api/generated/models/replaceSavedSupportViewDraftDto.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Retenive Backend Platform
 * CMS, integration, chat and realtime API for Retenive AI Assistant
 * OpenAPI spec version: 0.1.0
 */
import type { ReplaceSavedSupportViewDraftDtoDraft } from "./replaceSavedSupportViewDraftDtoDraft";

export interface ReplaceSavedSupportViewDraftDto {
  draft: ReplaceSavedSupportViewDraftDtoDraft;
}